UI/team_22_interiit_2024/backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.0
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
auth0-python==4.7.2
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.12.3
billiard==4.2.1
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cricketstats==0.1.6
cryptography==43.0.3
daphne==4.1.2
datadog-api-client==2.12.0
Deprecated==1.2.14
diff-match-patch==20241021
dill==0.3.7
distlib==0.3.9
distro==1.9.0
Django==5.1.3
django-cors-headers==4.6.0
django-environ==0.11.2
django-filter==24.3
django-import-export==4.2.0
djangochannelsrestframework==1.2.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
fal_client==0.5.6
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
fuzzywuzzy==0.18.0
google==3.0.0
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.153.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
greenlet==3.1.1
groq==0.12.0
grpc-interceptor==0.15.4
grpcio==1.64.0
gTTS==2.5.4
gunicorn==23.0.0
h11==0.14.0
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
httpx-sse==0.4.0
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.5.0
incremental==24.7.2
inflection==0.5.1
isolate==0.14.3
isolate-proto==0.0.37
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy==2.1.3
opentelemetry-api==1.28.1
opentelemetry-sdk==1.28.1
opentelemetry-semantic-conventions==0.49b1
packaging==24.2
pandas==2.2.3
pathspec==0.11.2
platformdirs==4.3.6
portalocker==2.10.1
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
py-expression-eval==0.3.14
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==1.10.19
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pyOpenSSL==24.2.1
pyparsing==3.2.0
PyYAML==6.0.2
redis==5.2.0
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
service-identity==24.2.0
setuptools==75.5.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
sqlparse==0.5.1
structlog==22.3.0
sympy==1.13.1
tablib==3.7.0
tblib==3.0.0
termcolor==2.5.0
threadpoolctl==3.5.0
torch==2.5.1
tqdm==4.67.0
Twisted==24.10.0
txaio==23.1.1
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
virtualenv==20.27.1
wcwidth==0.2.13
wrapt==1.16.0
yarl==1.17.1
zipp==3.21.0
py_expression_eval==0.3.14
